# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Incidents API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1"
    },
    "basePath": "/",
    "paths": {
        "/incidents": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "list incidents",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "incidents"
                ],
                "summary": "list incidents",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "limits size of result; default 100",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "offset to be used in combination with limit, default 0",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "default id.asc, sortable by id, external_task_id, process_instance_id, process_definition_id, time",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "filter by process_definition_id",
                        "name": "process_definition_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "filter by process_instance_id",
                        "name": "process_instance_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "filter by external_task_id",
                        "name": "external_task_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/messages.IncidentMessage"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "create incident, user must be admin",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "incidents"
                ],
                "summary": "create incident",
                "parameters": [
                    {
                        "description": "Incident",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/messages.IncidentMessage"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/incidents/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "get incident",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "incidents"
                ],
                "summary": "get incident",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Incident Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/messages.IncidentMessage"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/on-incident-handler": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "set on incident handler, user must be admin",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "incidents"
                ],
                "summary": "set on incident handler",
                "parameters": [
                    {
                        "description": "Incident-Handler",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/messages.OnIncident"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/process-definitions/{id}": {
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "delete incidents by process-definition id, user must be admin",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "incidents"
                ],
                "summary": "delete incidents by process-definition id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "process-definition id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/process-instances/{id}": {
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "delete incidents by process-instance id, user must be admin",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "incidents"
                ],
                "summary": "delete incidents by process-instance id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "process-instance id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "messages.IncidentMessage": {
            "type": "object",
            "properties": {
                "deployment_name": {
                    "type": "string"
                },
                "error_message": {
                    "type": "string"
                },
                "external_task_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "msg_version": {
                    "description": "from version 3 onward will be set in KafkaIncidentsCommand and be copied to this field",
                    "type": "integer"
                },
                "process_definition_id": {
                    "type": "string"
                },
                "process_instance_id": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                },
                "worker_id": {
                    "type": "string"
                }
            }
        },
        "messages.OnIncident": {
            "type": "object",
            "properties": {
                "notify": {
                    "type": "boolean"
                },
                "process_definition_id": {
                    "type": "string"
                },
                "restart": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}